refactor: move helper function to new folder
